Adding raw transaction relay to EVM: allows to send Ethereum signed transactions that will get executed on behalf of the sender

+++ chain/chain/src/types.rs
@@ -580,7 +580,7 @@ pub trait RuntimeAdapter: Send + Sync {
     ) -> Result<bool, Error>;
 
     #[cfg(feature = "protocol_feature_evm")]
-    fn evm_chain_id(&self) -> u128;
+    fn evm_chain_id(&self) -> u64;
 
     /// Build receipts hashes.
     // Due to borsh serialization constraints, we have to use `&Vec<Receipt>` instead of `&[Receipt]`
